Downloads/ManON/requirements.txt
amqp==5.0.9
asgiref==3.5.0
astroid==2.8.6
asttokens==2.0.5
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-aliases==1.0.1
click-default-group==1.2.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
DateTime==4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-doctor==2.6.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dparse==0.5.1
Faker==12.3.0
generics==3.4.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
kombu==5.2.3
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.11.1
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.3
requests==2.27.1
rich==11.2.0
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
zope.interface==5.4.0
